+++ src/services/api/mappers/tokens/toUpdateTokenRequest.ts
@@ -1,8 +1,10 @@
 import { UpdateToken } from '../../../../types';
 import { convertToSnakeCase } from '../case-conversion-utils';
-import { isUpdateToken } from './type-guards';
+import { isUpdateToken, UpdateTokenResponse } from './type-guards';
 
-export const toUpdateTokenRequest = (payload: Partial<UpdateToken>) => {
+export const toUpdateTokenRequest = (
+  payload: Partial<UpdateToken>
+): UpdateTokenResponse | undefined => {
   const _payload = convertToSnakeCase(payload);
 
   return isUpdateToken(_payload)